refactor discard module to facilitate hunk-based implementation.
